Correccion de home(yoli)
Correccion de imagenes en distintos dispocitivos
Se agregaron imagenes.
Textos
