+++ src/app/app.component.html
@@ -5,7 +5,7 @@
     <ion-header class="ion-no-border">
       <div class="bg"></div>
       <ion-item lines="none">
-      <ion-avatar slot="start">
+      <ion-avatar slot="start" class="avbase">
       <img src="../assets/img/Ax.jpeg" alt="Axolotl">
       </ion-avatar>
       <ion-label color="light">
