+++ src/app/app.component.scss
@@ -17,7 +17,7 @@ ion-Header {
         position: absolute;
         top: 1rem;
         left: 0;
-        ion-avatar{
+        .avbase{
             height: 5rem;
             width: 5rem;
             //border-radius: 50%;
