+++ src/app/componentes/header/header.component.scss
@@ -1,0 +1,21 @@
+ion-toolbar{
+    --background: linear-gradient(135deg, rgb(10, 182, 18) 30%, rgb(233, 236, 20));   
+    .title{
+        padding-left: 1px;
+    }
+}
+
+
+
+
+// ion-toolbar {
+//   --color: white;
+
+//   --border-color: #f24aec;
+//   --border-width: 4px 0;
+//   --border-style: double;
+
+//   --min-height: 80px;
+//   --padding-top: 20px;
+//   --padding-bottom: 20px;
+// }
