+++ src/app/pages/controles/controles.page.html
@@ -1,20 +1,28 @@
-<ion-header [translucent]="true">
-  <ion-toolbar>
-    <!-- MENU -->
-    <ion-buttons slot="start">
-      <ion-menu-button menu="first"></ion-menu-button>
-    </ion-buttons>
-    <!-- MENU -->
-    <ion-title>Controles</ion-title>
-  </ion-toolbar>
-</ion-header>
+<app-header titulo="Controles"></app-header>
 
-<ion-content [fullscreen]="true">
-  <ion-header collapse="condense">
-    <ion-toolbar>
-      <ion-title size="large">Controles</ion-title>
-    </ion-toolbar>
-  </ion-header>
-    <!-- Aqui va el contenido -->
-  <h1>Contents</h1>
+<ion-content fullscreen="true" class="ion-padding">
+  <!-- Contenido -->
+  <div class="space"></div>
+  
+  <ion-grid class="image-grid">
+  <ion-row class="ion-justify-content-center ion-align-items-center">
+    <div class="image-wrapper">
+      <ion-avatar class="Ax">
+        <ion-img src="../../../assets/img/Ax.jpeg"></ion-img>
+      </ion-avatar>
+      <ion-avatar class="lhc">
+        <ion-img src="../../../assets/img/yoliuno.png"></ion-img>
+      </ion-avatar>
+    </div>
+  </ion-row>
+</ion-grid>
+
+
+
+
+
+
+
+
+  <!-- Contenido -->
 </ion-content>
