+++ src/app/pages/items/items.page.html
@@ -1,21 +1,7 @@
-<ion-header [translucent]="true">
-  <ion-toolbar>
-    <!-- MENU -->
-    <ion-buttons slot="start">
-      <ion-menu-button menu="first"></ion-menu-button>
-    </ion-buttons>
-    <!-- MENU -->
-    <ion-title>Items</ion-title>
-  </ion-toolbar>
-</ion-header>
+<app-header titulo="Items"></app-header>
 
-
-<ion-content [fullscreen]="true">
-  <ion-header collapse="condense">
-    <ion-toolbar>
-      <ion-title size="large">Items</ion-title>
-    </ion-toolbar>
-  </ion-header>
-  <!-- Aqui va el contenido -->
-  <h1>Contents</h1>
+<ion-content fullscreen="true" class="ion-padding">
+  <!-- Contenido -->
+  <h1>Contenido</h1>
+  <!-- Contenido -->
 </ion-content>
