+++ src/app/pages/mercancia/mercancia.page.html
@@ -1,20 +1,8 @@
-<ion-header [translucent]="true">
-  <ion-toolbar>
-    <!-- MENU -->
-    <ion-buttons slot="start">
-      <ion-menu-button menu="first"></ion-menu-button>
-    </ion-buttons>
-    <!-- MENU -->
-    <ion-title>Mercancia</ion-title>
-  </ion-toolbar>
-</ion-header>
+<app-header titulo="Mercancia"></app-header>
 
-<ion-content [fullscreen]="true">
-  <ion-header collapse="condense">
-    <ion-toolbar>
-      <ion-title size="large">Mercancia</ion-title>
-    </ion-toolbar>
-  </ion-header>
-    <!-- Aqui va el contenido -->
-  <h1>Contents</h1>
+<ion-content fullscreen="true" class="ion-padding">
+  <!-- Contenido -->
+  <h1>Contenido</h1>
+  <!-- Contenido -->
 </ion-content>
+
