+++ src/app/pages/yoli/yoli.page.html
@@ -1,35 +1,40 @@
-<ion-header [translucent]="true">
-  <ion-toolbar>
-    <!-- MENU -->
-    <ion-buttons slot="start">
-      <ion-menu-button menu="first"></ion-menu-button>
-    </ion-buttons>
-    <!-- MENU -->
-    <ion-title>Yoli</ion-title>
-  </ion-toolbar>
-</ion-header>
+<app-header titulo="Yoli"></app-header>
 
-<ion-content [fullscreen]="true">
-  <ion-header collapse="condense">
-    <ion-toolbar>
-      <ion-title size="large">Yoli</ion-title>
-    </ion-toolbar>
-  </ion-header>
-    <!-- Aqui va el contenido -->
+  <ion-content [fullscreen]="true">
+    <!-- Contenido -->
+    <div class="space"></div>
+  
+    <ion-grid class="image-grid">
+      <ion-row class="ion-justify-content-center ion-align-items-center">
+        <div class="image-wrapper">
+          <ion-avatar class="Ax">
+            <ion-img src="../../../assets/img/Ax.jpeg"></ion-img>
+          </ion-avatar>
+          <ion-avatar class="lhc">
+            <ion-img src="../../../assets/img/yoliuno.png"></ion-img>
+          </ion-avatar>
+        </div>
+      </ion-row>
+    </ion-grid>
 
-        <ion-content>
-      <ion-card *ngFor="let personaje of personajes">
-        <img [src]="personaje.data.imagen" style="height: 40hv; width: 60hv;"/>
-        <ion-card-header>
-          <ion-card-title>Nombre: {{personaje.data.nombre}}</ion-card-title>
-          <ion-card-subtitle>Arma principal: {{personaje.data.armabase}}</ion-card-subtitle>
-          <ion-card-subtitle>Arma secundaria: {{personaje.data.armasecundaria}}</ion-card-subtitle>
-        </ion-card-header>
-        <ion-card-content>
-          <h3>Descripcion: {{personaje.data.descripcion}}</h3>
-        </ion-card-content>
-      </ion-card>
-    </ion-content>
-    <!-- Aqui va el contenido -->
- 
-</ion-content>
+    <ion-grid>
+      <ion-row>
+        <ion-col class="text">
+          <ion-text> "Text" Lorem ipsum dolor sit amet consectetur adipisicing elit. Eius suscipit voluptates ea, porro, voluptatibus nesciunt aliquid doloremque ad eos placeat odio excepturi provident praesentium dolore voluptate. Voluptas reprehenderit reiciendis doloribus!</ion-text>
+        </ion-col>
+      </ion-row>
+      <ion-row class="row-g">
+        <ion-img class="img" src="../../../assets/img/gerrero.jpeg"></ion-img>
+      </ion-row>
+      <ion-row class="text">
+          <ion-text> "Text" Lorem ipsum dolor sit amet consectetur adipisicing elit. Eius suscipit voluptates ea, porro, voluptatibus nesciunt aliquid doloremque ad eos placeat odio excepturi provident praesentium dolore voluptate. Voluptas reprehenderit reiciendis doloribus!</ion-text>
+      </ion-row>
+    </ion-grid>
+
+
+
+
+  </ion-content>
+  <!-- aqui va el contenido -->
+
+
