+++ src/app/pages/yoli/yoli.page.ts
@@ -1,12 +1,5 @@
 import { Component, OnInit } from '@angular/core';
 
-//Importar el sevicio
-import { FirebaseService } from 'src/app/services/firebase.service';
-//Importar las intercaes
-import { Personajes, Controles, Items, Mercancia, Datos } from 'src/app/interfaces/interfaces';
-
-
-
 @Component({
   selector: 'app-yoli',
   templateUrl: './yoli.page.html',
@@ -14,26 +7,9 @@ import { Personajes, Controles, Items, Mercancia, Datos } from 'src/app/interfac
 })
 export class YoliPage implements OnInit {
 
-  personajes: Datos[] = [];
-
-  //inicializar el objeto
-  constructor(private servicioFirebase: FirebaseService) { }
-
-
-  async verDetalle(id: string) {
-
-  }
+  constructor() { }
 
   ngOnInit() {
-    this.servicioFirebase.getDatos().subscribe((response) => {
-      console.log(response[0].payload.doc)
-      response.forEach(obj => {
-        this.personajes.push({
-          id: obj.payload.doc.id,
-          data: <Personajes>obj.payload.doc.data(),
-        });
-      });
-    });
   }
 
 }
